+++ keycloak/README.adoc
@@ -1,6 +1,10 @@
 = Keycloak Project
-:hardbreaks:
 
+:toc: left
+:icons: font
+:source-highlighter: rouge
+:description: Example project for using Hibernate Reactive Panache in Quarkus.
+:hardbreaks:
 
 == Set Up
 
@@ -15,7 +19,6 @@ with yor firewall.
 ----
 firewall-cmd --add-port 8090/tcp --permanent
 firewall-cmd --add-port 8043/tcp --permanent
-
 ----
 
 
@@ -46,27 +49,116 @@ Firstly it is necessary to copy the realm.json in the tmp directory, as this dir
 Beware: You need to be located in your root project directory in order to perform the following command.
 [source,bash]
 ----
-cp src/test/resources/trikora-realm.json /tmp
+mkdir /tmp/keycloak_tmp/
+cp src/test/resources/trikora-realm.json /tmp/keycloak_tmp/
 ----
 Then we can execute:
 
 [source,bash]
 ----
-podman run --name trikora_keycloak_quarkus_example --security-opt label=disable --rm --pod trikora_keycloak_quarkus_example_pod -e KEYCLOAK_USER=admin -e KEYCLOAK_PASSWORD=admin -e KEYCLOAK_IMPORT=/tmp/trikora-realm.json -v /tmp/trikora-realm.json:/tmp/trikora-realm.json quay.io/keycloak/keycloak:15.0.2
+podman run --name trikora_keycloak_quarkus_example --security-opt label=disable \
+  --rm --pod trikora_keycloak_quarkus_example_pod \
+  -e KEYCLOAK_USER=admin -e KEYCLOAK_PASSWORD=admin -e KEYCLOAK_IMPORT=/tmp/trikora-realm-old.json \
+  -v /tmp/keycloak_tmp/:/tmp/ \
+  quay.io/keycloak/keycloak:15.0.2
 ----
 
 NOTE: the --rm flag is set for deleting the pod after the execution. You can kill the process with `^C`.
 
-== Actions in the keycloak console.
+== Actions in the keycloak console
 
 * Link to keycloak console: https://localhost:8543/auth/admin/master/console/#/realms/quarkus
 
-== Export realm json file.
-TODO
+==== Roles
+* We can access the roles section by the left navigator in the main page or clicking here:
+https://localhost:8543/auth/admin/master/console/#/realms/quarkus/roles
+
+* In this section we can add roles to our application that will be saved in the realms.json that we have located in the tmp directory.
+
+==== User Management
+
+* If the user section, we can edit all the information regarding the user such as the email, password ...
+but indeed we have the possibility of assigning new roles to the user as well as enroll them to different groups.
+
+`TODO` Create permission add it to a role and then assign it to a group of users
+
+== Security Management
+* In this section we will describe how keycloak filter the URLs with the proper permissions.
+
+* Firstly, we have to locate in Clients -> backend-service -> Authorization -> Settings -> Resources.
+Link: https://localhost:8543/auth/admin/master/console/#/realms/quarkus/clients/0ac5df91-e044-4051-bd03-106a3a5fb9cc/authz/resource-server/resource
+
+* Then, we have to create a new resource where we will establish the URLs that are going to be restricted. For doing so, there is a button `Create` in the upper-right side of the main layout.
+
+* After that, it is necessary to add an Associated Permission to the Resource, now we have to decide who is going to have access to the service by creating a new policy. A policy, may be a Role, Group, Client scope or even a single user. Once we have linked the policy to the resource, just the desired users could access to the services.
+
+* If a user without the proper permissions tries to access the service, keycloak will return a http response with status 403 (FORBIDDEN).
+
+== Backup and Restore
+* As the keycloak admin console does not allow exporting users, we highly recommend using the command line solution in order to fully export your project.
+
+=== Export
+* In this section we will describe how to export a realm.json in a single file from a running container using podman.
+
+[source,shell script]
+----
+[podman | docker] exec -it <pod_name> opt/jboss/keycloak/bin/standalone.sh
+        -Djboss.socket.binding.port-offset=100
+        -Dkeycloak.migration.action=[export | import]
+        -Dkeycloak.migration.provider=[singleFile | dir]
+        -Dkeycloak.migration.dir=<DIR TO EXPORT TO> Use only iff .migration.provider=dir
+        -Dkeycloak.migration.realmName=<REALM_NAME_TO_EXPORT>
+        -Dkeycloak.migration.usersExportStrategy=[DIFFERENT_FILES | SKIP | REALM_FILE | SAME_FILE]
+        -Dkeycloak.migration.usersPerFile=<integer_value> Use only iff .usersExportStrategy=DIFFERENT_FILES
+        -Dkeycloak.migration.file=<FILE TO EXPORT TO>
+----
+
+
+[source,bash]
+----
+podman exec -it trikora_keycloak_quarkus_example /opt/jboss/keycloak/bin/standalone.sh \
+ -Djboss.socket.binding.port-offset=100 \
+ -Dkeycloak.migration.action=export \
+ -Dkeycloak.migration.provider=singleFile \
+ -Dkeycloak.migration.realmName=trikorasolutions \
+ -Dkeycloak.migration.usersExportStrategy=REALM_FILE \
+ -Dkeycloak.migration.file=/tmp/my_realm.json
+----
+
+=== Import
+
+[source,bash]
+----
+podman exec -it trikora_keycloak_quarkus_example /opt/jboss/keycloak/bin/standalone.sh \
+ -Djboss.socket.binding.port-offset=100 \
+ -Dkeycloak.migration.action=import \
+ -Dkeycloak.migration.provider=singleFile \
+ -Dkeycloak.migration.realmName=quarkus \
+ -Dkeycloak.migration.usersExportStrategy=REALM_FILE \
+ -Dkeycloak.migration.file=/tmp/my_realm.json \
+ -Dkeycloak.migration.strategy=OVERWRITE_EXISTING
+----
+
+NOTE:  [podman | docker] exec -it <container_name> <PATH_TO_KEYCLOAK_IN_THE_POD>/bin/standalone.sh
+ -Djboss.socket.binding.port-offset=100
+ -Dkeycloak.migration.action=import
+ -Dkeycloak.migration.provider=singleFile
+ -Dkeycloak.migration.realmName=quarkus
+ -Dkeycloak.migration.usersExportStrategy=REALM_FILE
+ -Dkeycloak.migration.file=<FILE_TO_IMPORT>
+ -Dkeycloak.migration.strategy=[OVERWRITE_EXISTING | IGNORE_EXISTING]
+
+== Import several files in a single realm
+* If we have stored the data of the project split in several files, we can merge it in a single project just by importing the files as they are a list separated by commas:
+-Dkeycloak.import=/tmp/realm1.json,/tmp/realm2.json
+
+WARNING: You cannot use the keycloak.import parameter with keycloak.migration.X parameters.
+If you use these parameters together, keycloak.import parameter will be ignored. The keycloak.import mechanism ignores the realms which already exist in the project.
+The keycloak.import mechanism is convenient for development purposes, but if more flexibility is needed, use the keycloak.migration.X parameters.
 
 == Troubleshooting
 
-=== Permission denied
+=== Permission denied when running the piod
 
 *Problem*
