+++ keycloak/src/test/java/com/trikorasolutions/example/AdminResourceTest.java
@@ -5,14 +5,13 @@ import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.Test;
 import org.keycloak.representations.AccessTokenResponse;
+
+import static com.trikorasolutions.example.KeycloakInfo.getAccessToken;
 import static javax.ws.rs.core.Response.Status.*;
 
 
 @QuarkusTest
-public class PolicyEnforcerTest {
-
-    private static final String KEYCLOAK_SERVER_URL = System.getProperty("keycloak.url", "https://localhost:8543/auth");
-    private static final String KEYCLOAK_REALM = "quarkus";
+public class AdminResourceTest {
 
     static {
         RestAssured.useRelaxedHTTPSValidation();
@@ -46,24 +45,4 @@ public class PolicyEnforcerTest {
                 .statusCode(OK.getStatusCode());
     }
 
-    @Test
-    public void testPublicResource() {
-        RestAssured.given()
-                .when().get("/api/public")
-                .then()
-                .statusCode(NO_CONTENT.getStatusCode());
-    }
-
-    private String getAccessToken(String userName) {
-        return RestAssured
-                .given()
-                .param("grant_type", "password")
-                .param("username", userName)
-                .param("password", userName)
-                .param("client_id", "backend-service")
-                .param("client_secret", "secret")
-                .when()
-                .post(KEYCLOAK_SERVER_URL + "/realms/" + KEYCLOAK_REALM + "/protocol/openid-connect/token")
-                .as(AccessTokenResponse.class).getToken();
-    }
 }
